% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scraping.R
\name{data_scraping}
\alias{data_scraping}
\title{Read in data on all available graduating classes}
\usage{
data_scraping()
}
\value{
A dataframe object containing information about all available graduating classes
}
\description{
Read in data on all available graduating classes. The information of interest
             includes student names, latin honors, department honors, Clark fellowship,
             Phi Kappa Beta membership and Sigma Xi membership.
}
\keyword{StudentInfo}
\keyword{readStudent}

