% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_info.R
\docType{package}
\name{gradinfo}
\alias{gradinfo}
\alias{gradinfo-package}
\title{gradinfo: access and analyze data about Williams College graduating classes}
\description{
gradinfo reads in information about graduating classes of Williams College from 2001 to 2016 and
provide summary statistics on the data.
This package has three functions"

\itemize{
\item readStudentInfo: reads in information about graduating classes of Williams College from locally stored
text files and return a data frame containing 9 variables. For more information, type ?readStudentInfo
\item statsummary: provides summary statistics of the subject selected by the user in either data frame or graphical
format. Check ?statsummary for more information
\item data_scraping: reads in information from all locally stored text files and creates a dataframe from the information
read in. The new dataframe will be saved in the database of this package.
}
}

