% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R, R/majors.R
\docType{data}
\name{majors}
\alias{majors}
\title{Data on majors of Williams College's graduating classes from 2007 to 2016}
\format{A data frame with 9 variables
\itemize{
  \item \strong{Majors}: Majors of graduating students in abbreviated format
  \item \strong{Year}: Graduating years, ranging from 2007 to 2016
  \item \strong{Number.of.Students}: Number of students pursuing a particular major
                                     in a specificed graduating class
  \item \strong{Percentage}: The percentage equivalent of the value in \strong{Number.of.Students} column
  }}
\description{
This dataframe contains information about the majors of Williams College's graduating classes from 2007
to 2016.

This dataframe contains information about the numbers of students in each major at Williams College for each year 
from 2007 to 2016
}
\keyword{data,}
\keyword{graduate,}
\keyword{major}
\keyword{major,}
\keyword{majors}
\keyword{majors,}
\keyword{undergrad,}
\keyword{williams,}

