% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStudentInfo.R
\name{readStudentInfo}
\alias{readStudentInfo}
\title{Read in data on a specified graduating class}
\usage{
readStudentInfo(data, grad.year)
}
\arguments{
\item{data}{Name of the source text file containing information about the interested
graduating class. The argument should follow this format "yyyy-yy.txt".
For instance,if the interested graduating year is 2001, the name of the source
text file would be "2000-01.txt". Accepted graduating year ranges from 2001 to
2016}

\item{grad.year}{Sets the year in which the given data was collected. The argument should
be a 4 digit integer between 2001 and 2016 (inclusive)}
}
\value{
A dataframe object containing information about the specified graduating class
}
\description{
Read in information about the graduating class specified by the \code{grad.year}
             parameter. The information of interest includes student names, latin honors,
             department honors, Clark fellowship, Phi Kappa Beta membership and Sigma Xi
             membership
}
\examples{
readStudentInfo("2000-01.txt", 2001)
readStudentInfo("2001-02.txt", 2002)

}
\keyword{StudentInfo}
\keyword{readStudent}

