% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{statsummary}
\alias{statsummary}
\title{Summary Statistics}
\usage{
readStudentInfo(data, grad.year)
}
\arguments{
\item{type}{The subject of analysis. Possible options are \code{undergrad, grad, latin, department, gender}.
\itemize{
 \item \strong{undergrad}: generates summary statistics about undegraduate students
 \item \strong{grad}: generates summary statistics about graduate students
 \item \strong{latin}: generates summary statistics about the distribution of Latin honors at Williams
 \item \strong{department}: generates summary statistics about the distribution of department honors at
                            Williams
 \item \strong{gender}: generates summary statistics about the overall gender ratio and the gender ratios
                        of each department.
}}

\item{format}{This argument has three values \code("numeric", "timeseries", "distribution")
\itemize{
 \item \strong{numeric}: returns a data frame containing relevant summary statistics
 \item \strong{timeseries}: generates graphical summary that displays changes of the interested statistics over time.
 \item \strong{distribution}: generates graphical summary that displays distribution of the interested subject.
}}
}
\value{
A dataframe containing summary statistics if \code{format == "numeric"}
}
\description{
Uses data from the williams_grad dataset to compute and display appropriate summary
statistics in either tabular or graphical format.
}
\examples{
statsummary("undergrad", "numeric") # Return a dataframe containing the numbers of undergraduate students by years
statsummary("latin", "timeseries")  # Return a bar chart displaying the percentage of students receiving Summa Cum Laude,
                                    # Magna Cum Laude, Cum Laude and no Latin honors over time.
statsummay("depertment", "distribution") # Return a bar chart showing top 5 most popular department (ranked by percentage of total students
                                           majoring in the department)

}
\keyword{analysis,}
\keyword{data_analysis}

