% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_majors.R
\docType{data}
\name{sum_majors}
\alias{sum_majors}
\title{Data on majors at Williams College from 2007 to 2016}
\format{A data frame with 2 variables
\itemize{
  \item \strong{Year}: The year in which the number of major was recorded
  \item \strong{Number.of.Majors}: Number of majors pursued by students who graduated in a given year
  }}
\description{
This dataframe contains information about the numbers of students in each major at Williams College for each year 
from 2007 to 2016
}
\keyword{data,}
\keyword{graduate,}
\keyword{major,}
\keyword{majors}
\keyword{undergrad,}
\keyword{williams,}

